requirements.txt
pandas==2.0.3
numpy==1.24.3
yfinance==0.2.31
matplotlib==3.7.2
seaborn==0.13.0
scikit-learn==1.3.0
gspread==5.12.0
oauth2client==4.1.3
ta==0.10.2
python-telegram-bot==20.4
pytest==7.4.0
python-dotenv==1.0.0